Move tags logic from service to the new tags model class
