+++ packages/framework/tests/Feature/Services/PublicationServiceTest.php
@@ -148,7 +148,6 @@ class PublicationServiceTest extends TestCase
         $this->assertTrue(PublicationService::publicationTypeExists('test-publication'));
         $this->assertFalse(PublicationService::publicationTypeExists('foo'));
     }
-
     public function testGetAllTags()
     {
         $tags = [
@@ -159,11 +158,6 @@ class PublicationServiceTest extends TestCase
         ];
         $this->file('tags.json', json_encode($tags));
         $this->assertSame($tags, PublicationService::getAllTags()->toArray());
-    }
-
-    public function testGetAllTagsWithNoTags()
-    {
-        $this->assertSame([], PublicationService::getAllTags()->toArray());
     }
 
     public function testGetValuesForTagName()
@@ -182,25 +176,6 @@ class PublicationServiceTest extends TestCase
         $this->file('tags.json', json_encode($tags));
 
         $this->assertSame(['bar', 'baz'], PublicationService::getValuesForTagName('foo')->toArray());
-    }
-
-    public function testGetValuesForTagNameWithMissingTagName()
-    {
-        $tags = [
-            'foo' => [
-                'bar',
-                'baz',
-            ],
-        ];
-
-        $this->file('tags.json', json_encode($tags));
-
-        $this->assertSame([], PublicationService::getValuesForTagName('bar')->toArray());
-    }
-
-    public function testGetValuesForTagNameWithNoTags()
-    {
-        $this->assertSame([], PublicationService::getValuesForTagName('foo')->toArray());
     }
 
     protected function createPublicationType(): void
